src/utils.py
from pathlib import Path
import yaml

def load_config(config_path: str | Path = "configs/config.yaml") -> dict:
    with Path(config_path).open("r", encoding="utf-8") as f:
        return yaml.safe_load(f)

def ensure_dir(path: str | Path) -> Path:
    path = Path(path)
    path.mkdir(parents=True, exist_ok=True)
    return path
