requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.12.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0.1
jupyter>=1.0.0
notebook>=7.0.0
tqdm>=4.66.0
category-encoders>=2.6.0
lightgbm>=4.3.0
xgboost>=2.0.0
catboost>=1.2.0
statsmodels>=0.14.0
pyarrow>=15.0.0
joblib>=1.3.0
